+++ R/functions_data.R
@@ -21,6 +21,8 @@
 #' @param startTime a numeric `as.Date` format value; prior events are
 #' disregarded.
 #' @param ... Not further arguments are required.
+#' @param envir an `environment` where the nodes and linked events
+#'   objects are available.
 #' @return The respective object updated accordingly to the events link to it.
 #' For `nodes.goldfish` object the attributes are updated according to the
 #' events linked to them.
@@ -38,7 +40,8 @@
 #' bilatnet <- defineNetwork(bilatnet, nodes = states, directed = FALSE)
 #' bilatnet <- linkEvents(bilatnet, bilatchanges, nodes = states)
 #'
-#' updateStates <- as.data.frame(states,
+#' updateStates <- as.data.frame(
+#'   states,
 #'   time = as.numeric(as.POSIXct("1965-12-31"))
 #' )
 #'
@@ -52,8 +55,10 @@ NULL
 # Create a data frame from a dynamic nodes object
 #' @export
 #' @rdname update-method
-as.data.frame.nodes.goldfish <- function(x, ..., time = -Inf,
-                                         startTime = -Inf) {
+as.data.frame.nodes.goldfish <- function(
+    x, ..., time = -Inf,
+    startTime = -Inf, envir = new.env()
+) {
   df <- x
   dynamicAttributes <- attr(df, "dynamicAttribute")
   eventNames <- attr(df, "events")
@@ -64,8 +69,8 @@ as.data.frame.nodes.goldfish <- function(x, ..., time = -Inf,
     return(df)
   }
   for (i in seq_along(eventNames)) {
-    events <- get(eventNames[i])
-    events <- sanitizeEvents(events, df)
+    events <- get(eventNames[i], envir = envir)
+    events <- sanitizeEvents(events, df, envir = envir)
     events <- events[events$time >= startTime & events$time < time, ]
 
     if (nrow(events) > 0 && !is.null(events$replace)) {