+++ NEWS.md
@@ -1,7 +1,13 @@
+# goldfish 1.6.6
+
+* Debugging `GatherPreprocessing()`.
+
 # goldfish 1.6.5
 
 * Solves `startTime` and `endTime` bug on `DyNAM` and `REM` models
   preprocessing.
+* Export `GatherPreprocessing()`. Experimental functionality.
+* Introduces experimental functionality `simulation()`.
 * Clean unnecessary functions imports.
 * Solves `aes_string()` deprecation.
 * Solves issue on `C++` engine on DyNAM-rate.